requirements.txt
Django==2.1.7
django-environ==0.4.5
djangorestframework==3.9.1
djangorestframework-simplejwt==3.3
pep8==1.7.1
PyJWT==1.7.1
pytz==2018.9
